added base bundle structure, moved selector files to new bundle
