+++ bundles/AdminBundle/templates/admin/index/index.html.twig
@@ -309,10 +309,6 @@
 
     "pimcore/element/abstract.js",
     "pimcore/element/abstractPreview.js",
-    "pimcore/element/selector/selector.js",
-    "pimcore/element/selector/abstract.js",
-    "pimcore/element/selector/document.js",
-    "pimcore/element/selector/asset.js",
     "pimcore/element/properties.js",
     "pimcore/element/scheduler.js",
     "pimcore/element/dependencies.js",
@@ -339,7 +335,6 @@
     "pimcore/object/helpers/optionEditor.js",
     "pimcore/object/helpers/imageGalleryDropZone.js",
     "pimcore/object/helpers/imageGalleryPanel.js",
-    "pimcore/element/selector/object.js",
     "pimcore/element/tag/configuration.js",
     "pimcore/element/tag/assignment.js",
     "pimcore/element/tag/tree.js",
