+++ bundles/SimpleBackendSearchBundle/public/js/pimcore/selector/abstract.js
