+++ bundles/SimpleBackendSearchBundle/public/js/pimcore/selector/asset.js
