+++ bundles/SimpleBackendSearchBundle/public/js/pimcore/selector/document.js
