+++ bundles/SimpleBackendSearchBundle/public/js/pimcore/selector/object.js
