+++ bundles/SimpleBackendSearchBundle/public/js/pimcore/selector/selector.js
